GCM Logik in der SettingsActivity umgezogen
