+++ app/src/main/AndroidManifest.xml
@@ -29,7 +29,7 @@
         </activity>
 
         <receiver
-            android:name=".GCM.GcmBroadcastReceiver"
+            android:name=".gcm.GcmBroadcastReceiver"
             android:permission="com.google.android.c2dm.permission.SEND" >
             <intent-filter>
                 <action android:name="com.google.android.c2dm.intent.RECEIVE" />
@@ -38,7 +38,7 @@
             </intent-filter>
         </receiver>
 
-        <service android:name=".GCM.GcmIntentService" />
+        <service android:name=".gcm.GcmIntentService" />
 
         <activity
             android:name=".settings.SettingsActivity"
