+++ app/src/main/java/edu/kit/isco/kitalumniapp/gcm/GcmBroadcastReceiver.java
@@ -1,4 +1,4 @@
-package edu.kit.isco.kitalumniapp.GCM;
+package edu.kit.isco.kitalumniapp.gcm;
 
 import android.app.Activity;
 import android.content.ComponentName;
